fix: dont retrieve variable value snapshots separately from the insertion step
